pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]

[project]
name = "radiokommunikasjon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hypothesis>=6.151.6",
    "matplotlib>=3.10.8",
    "numba>=0.63.1",
    "pyadi-iio>=0.0.17",
    "pyldpc>=0.7.9",
    "pysdr>=1.2",
    "scipy>=1.14",
    "sympy>=1.14.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
norecursedirs = [".git", ".tox", ".hypothesis", "build", "dist", ".venv", "vendor", "pluto"]

[tool.uv.extra-build-dependencies]
pyldpc = ["numpy"]

[tool.ruff]
line-length = 120
extend-exclude = ["vendor"]

[tool.ruff.lint.pylint]
max-args = 8
max-returns = 8
